+++ GeneralsMD/Code/GameEngine/Source/GameLogic/ScriptEngine/ScriptActions.cpp
@@ -6117,7 +6117,7 @@ void ScriptActions::doTeamEmoticon(const AsciiString& teamName, const AsciiStrin
 	theTeam->getTeamAsAIGroup(theGroup.Peek());
 #endif
 
-	Int frames = (Int)( duration * LOGICFRAMES_PER_SECOND );
+	Int frames = (Int)( duration * static_cast<float>(LOGICFRAMES_PER_SECOND) );
 	theGroup->groupSetEmoticon( emoticonName, frames );
 }
 
@@ -6130,7 +6130,7 @@ void ScriptActions::doNamedEmoticon(const AsciiString& unitName, const AsciiStri
 		Drawable *draw = obj->getDrawable();
 		if( draw )
 		{
-			Int frames = (Int)( duration * LOGICFRAMES_PER_SECOND );
+			Int frames = (Int)( duration * static_cast<float>(LOGICFRAMES_PER_SECOND) );
 			draw->setEmoticon( emoticonName, frames );
 		}
 	}
@@ -6194,7 +6194,7 @@ void ScriptActions::doOverrideHulkLifetime( Real seconds )
 	else
 	{
 		// Convert real seconds into frames.
-		Int frames = (Int)(seconds * LOGICFRAMES_PER_SECOND);
+		Int frames = (Int)(seconds * static_cast<float>(LOGICFRAMES_PER_SECOND));
 		TheGameLogic->setHulkMaxLifetimeOverride(frames);
 	}
 }
